Restore known params fixture parity
